Use size_t type for index instead of uint64_t.
This avoids

	warning C4244: 'argument' : conversion from 'uint64_t' to 'unsigned int',
	possible loss of data

when using MSVC and generally seems more reasonable.

+++ include/cpptoml.h
@@ -271,7 +271,7 @@ class toml_array : public toml_base
         return values_;
     }
 
-    std::shared_ptr<toml_base> at(uint64_t idx) const
+    std::shared_ptr<toml_base> at(size_t idx) const
     {
         return values_.at(idx);
     }
